requirements.txt
Django==3.1.5
django-ckeditor==6.0.0v
validate-email==1.3
django-crispy-forms==1.10.0
Pillow==8.1.0